moved server,client,protocol packages under networking folder
