+++ pkg/network/protocol/message.go
