+++ pkg/network/protocol/packet.go
